new-classifier-deployment/requirements.txt
gradio>=4.21.0
huggingface_hub>=0.23.0
torch
torchvision
numpy
pillow